Make latency independent for each link, controlled through outgoingPeers
A node's "outgoingPeers" parameter is now an object (as opposed to an array).

For each outgoing connection, there is a latency associated with it. By default, this number is equal to the node's global latency, however it can be tweaked to differ from the node's latency.

The "outgoingPeers" link latencies are the ones that are now used for broadcasting network packets.